requirements.txt
python-telegram-bot==21.9
apscheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
pytz==2024.2
